ODS: fix loss of precision by unwanted use of float instead of double (raised by MSVC)
git-svn-id: https://svn.osgeo.org/gdal/trunk@31685 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/ods/ods_formula_node.cpp
@@ -1095,7 +1095,7 @@ int ods_formula_node::EvaluateBinaryArithmetic(IODSCellEvaluator* poEvaluator)
 
         if (papoSubExpr[1]->field_type == ODS_FIELD_TYPE_FLOAT)
         {
-            float dfVal;
+            double dfVal;
             switch (eOp)
             {
                 case ODS_ADD      : dfVal = (papoSubExpr[0]->float_value + papoSubExpr[1]->float_value); break;
